Allow a bigger precision in numbers and add relevant tests.

+++ RainLisp/Tokenization/NumberTokenizer.cs
@@ -8,7 +8,7 @@ namespace RainLisp.Tokenization
         private double _number;
         private bool _inFraction;
         private bool _negative;
-        private int _fractionDenominator;
+        private double _fractionDenominator;
 
         public NumberTokenizer() => Clear();
